Update licensing to be The maxCPUID Contributors
Currently, the copyright is held by Chris Blume.
This is correct for the time being. But it is not inviting
to others who may want to contribute to the project.

Instead, create an AUTHORS file where contributors can
add themselves and change the copyright ownership to
The maxCPUID Contributors.

Issue #10